# apps/bff_hdd/src/modules/shipment/shipment.controller.ts
import { Body, Controller, Get, Headers, HttpCode, Param, Post, Req } from '@nestjs/common';
import type { Request } from 'express';
import { AuthClaims } from '../../security/auth-claims';
import { DeliveryRequestDto } from './dto/delivery-request.dto';
import { ExceptionRequestDto } from './dto/exception-request.dto';
import { ShipmentResponseDto } from './dto/shipment-response.dto';
import { ShipmentService } from './shipment.service';
import { NoStoreResponse } from '../../security/no-store-response.decorator';

@NoStoreResponse()
@Controller('shipments')
export class ShipmentController {
  constructor(private readonly shipmentService: ShipmentService) {}

  @Get(':trackingNo')
  getShipment(@Param('trackingNo') trackingNo: string): Promise<ShipmentResponseDto> {
    return this.shipmentService.getShipment(trackingNo);
  }

  @Post(':trackingNo/delivery')
  @HttpCode(200)
  submitDelivery(
    @Req() request: Request,
    @Param('trackingNo') trackingNo: string,
    @Body() dto: DeliveryRequestDto,
    @Headers('x-idempotency-key') idempotencyKey?: string
  ): Promise<{ ok: boolean }> {
    const claims = (request as Request & { user: AuthClaims }).user;
    return this.shipmentService.submitDelivery(trackingNo, dto, claims, idempotencyKey);
  }

  @Post(':trackingNo/exception')
  @HttpCode(200)
  submitException(
    @Req() request: Request,
    @Param('trackingNo') trackingNo: string,
    @Body() dto: ExceptionRequestDto,
    @Headers('x-idempotency-key') idempotencyKey?: string
  ): Promise<{ ok: boolean }> {
    const claims = (request as Request & { user: AuthClaims }).user;
    return this.shipmentService.submitException(trackingNo, dto, claims, idempotencyKey);
  }
}