# contracts/openapi/huoduoduo-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Huoduoduo Express BFF API
  version: 0.2.10
  description: Transitional REST API in front of legacy SOAP services (UAT baseline).
  x-doc-id: HDD-OPENAPI-CONTRACT
  x-owner: BFF Lead
  x-last-updated: 2026-03-04
  x-review-status: Draft for management review
  x-language-pair:
    zh-TW: docs/architecture/API_SPEC_OPENAPI_PLAN24.zh-TW.md
    en: docs/architecture/API_SPEC_OPENAPI_PLAN24.en.md
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login and issue access/refresh tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login success.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "422":
          $ref: '#/components/responses/LegacyBusinessError'
        "502":
          $ref: '#/components/responses/LegacyTransportError'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and issue new access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  maxLength: 1024
      responses:
        "200":
          description: Refresh success.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
  /auth/logout:
    post:
      tags: [Auth]
      security:
        - bearerAuth: []
      summary: Revoke refresh token.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  maxLength: 1024
      responses:
        "200":
          description: Logout success.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                type: object
                properties:
                  revoked:
                    type: boolean
                  subject:
                    type: string
                    maxLength: 64
...
  /push/register:
    post:
      tags: [Push]
      security:
        - bearerAuth: []
      summary: Register device push token to legacy backend.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPushTokenRequest'
      responses:
        "200":
          description: Registration accepted.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  registeredAt:
                    type: string
                    format: date-time
                    maxLength: 40
  /push/unregister:
    post:
      tags: [Push]
      security:
        - bearerAuth: []
      summary: Unregister device push token from legacy backend.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnregisterPushTokenRequest'
      responses:
        "200":
          description: Unregister accepted.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  unregisteredAt:
                    type: string
                    format: date-time
                    maxLength: 40
...
  /shipments/{trackingNo}:
    get:
      tags: [Shipments]
      security:
        - bearerAuth: []
      summary: Get shipment detail (GetShipment_elf fallback GetShipment).
      parameters:
        - name: trackingNo
          in: path
          required: true
          schema:
            type: string
            maxLength: 32
      responses:
        "200":
          description: Shipment detail payload.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentResponse'
        "422":
          $ref: '#/components/responses/LegacyBusinessError'
        "502":
          $ref: '#/components/responses/LegacyTransportError'
...
  /shipments/{trackingNo}/exception:
    post:
      tags: [Shipments]
      security:
        - bearerAuth: []
      summary: Submit shipment exception payload.
      parameters:
        - name: trackingNo
          in: path
          required: true
          schema:
            type: string
            maxLength: 32
        - $ref: '#/components/parameters/IdempotencyKeyRequiredHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExceptionRequest'
      responses:
        "200":
          description: Accepted.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
...
  /reservations:
    get:
      tags: [Reservations]
      security:
        - bearerAuth: []
      summary: List reservations by mode.
      parameters:
        - $ref: '#/components/parameters/ReservationMode'
      responses:
        "200":
          description: Reservation list.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationResponse'
    post:
      tags: [Reservations]
      security:
        - bearerAuth: []
      summary: Create reservation by mode.
      parameters:
        - $ref: '#/components/parameters/ReservationMode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
      responses:
        "200":
          description: Reservation created.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                type: object
                properties:
                  reservationNo:
                    type: string
                    maxLength: 64
                  mode:
                    type: string
                    maxLength: 16
                    enum: [standard, bulk]
  /reservations/zip-areas:
    get:
      tags: [Reservations]
      security:
        - bearerAuth: []
      summary: Get reservable ZIP/area list.
      responses:
        "200":
          description: Reservable ZIP/area list.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationSupportListResponse'
  /reservations/available:
    get:
      tags: [Reservations]
      security:
        - bearerAuth: []
      summary: Get reservable shipment list (standard) by ZIP.
      parameters:
        - $ref: '#/components/parameters/ReservationSupportZip'
      responses:
        "200":
          description: Reservable standard shipment list.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationSupportListResponse'
  /reservations/available/bulk:
    get:
      tags: [Reservations]
      security:
        - bearerAuth: []
      summary: Get reservable shipment list (bulk) by ZIP.
      parameters:
        - $ref: '#/components/parameters/ReservationSupportZip'
      responses:
        "200":
          description: Reservable bulk shipment list.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationSupportListResponse'
  /reservations/area-codes:
    get:
      tags: [Reservations]
      security:
        - bearerAuth: []
      summary: Get reservation area-code list.
      responses:
        "200":
          description: Reservation area-code list.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationSupportListResponse'
  /reservations/arrived:
    get:
      tags: [Reservations]
      security:
        - bearerAuth: []
      summary: Get arrived-to-warehouse shipment list.
      responses:
        "200":
          description: Arrived shipment list.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationSupportListResponse'
  /reservations/{id}:
    delete:
      tags: [Reservations]
      security:
        - bearerAuth: []
      summary: Delete reservation by id and mode.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            maxLength: 64
        - $ref: '#/components/parameters/ReservationMode'
        - name: address
          in: query
          required: true
          schema:
            type: string
            maxLength: 512
      responses:
        "200":
          description: Reservation deleted.
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
            Pragma:
              $ref: '#/components/headers/PragmaNoCache'
            Expires:
              $ref: '#/components/headers/ExpiresZero'
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
...
components:
  headers:
    CacheControlNoStore:
      description: No-store policy for sensitive and transaction responses.
      schema:
        type: string
      example: no-store, no-cache, must-revalidate
    PragmaNoCache:
      description: Legacy-compatible no-cache directive.
      schema:
        type: string
      example: no-cache
    ExpiresZero:
      description: Expiration forced to immediate.
      schema:
        type: string
      example: "0"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdempotencyKeyRequiredHeader:
      name: X-Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        maxLength: 128
    ReservationMode:
      name: mode
      in: query
      required: false
      schema:
        type: string
        enum: [standard, bulk]
        default: standard
    ReservationSupportZip:
      name: zip
      in: query
      required: true
      schema:
        type: string
        maxLength: 64
...
  responses:
    LegacyBusinessError:
      description: Legacy backend business error.
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Expires:
          $ref: '#/components/headers/ExpiresZero'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    LegacyTransportError:
      description: Legacy backend transport/parsing error.
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Expires:
          $ref: '#/components/headers/ExpiresZero'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          maxLength: 64
        message:
          type: string
          maxLength: 1024
    LoginRequest:
      type: object
      required: [account, password, deviceId, platform]
      properties:
        account:
          type: string
          maxLength: 64
        password:
          type: string
          maxLength: 128
        deviceId:
          type: string
          maxLength: 64
        platform:
          type: string
          maxLength: 16
          enum: [android, ios]
    RefreshResponse:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
          maxLength: 4096
        refreshToken:
          type: string
          maxLength: 1024
    LoginResponse:
      type: object
      required: [accessToken, refreshToken, user, webviewBootstrap]
      properties:
        accessToken:
          type: string
          maxLength: 4096
        refreshToken:
          type: string
          maxLength: 1024
        user:
          $ref: '#/components/schemas/UserProfile'
        webviewBootstrap:
          $ref: '#/components/schemas/WebviewBootstrap'
    UserProfile:
      type: object
      required: [id, contractNo, name, role]
      properties:
        id:
          type: string
          maxLength: 64
        contractNo:
          type: string
          maxLength: 64
        name:
          type: string
          maxLength: 128
        role:
          type: string
          maxLength: 32
    WebviewBootstrap:
      type: object
      required: [baseUrl, registerUrl, resetPasswordUrl, cookies]
      properties:
        baseUrl:
          type: string
          format: uri
          maxLength: 2048
        registerUrl:
          type: string
          format: uri
          maxLength: 2048
        resetPasswordUrl:
          type: string
          format: uri
          maxLength: 2048
        cookies:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/WebCookie'
    WebCookie:
      type: object
      required: [name, value, domain, path, secure, httpOnly]
      properties:
        name:
          type: string
          maxLength: 64
        value:
          type: string
          maxLength: 4096
        domain:
          type: string
          maxLength: 255
        path:
          type: string
          maxLength: 255
        secure:
          type: boolean
        httpOnly:
          type: boolean
...
    ExceptionRequest:
      type: object
      required: [imageBase64, imageFileName, reasonCode, latitude, longitude]
      properties:
        driverId:
          type: string
          maxLength: 64
        imageBase64:
          type: string
        imageFileName:
          type: string
          maxLength: 255
        reasonCode:
          type: string
          maxLength: 64
        reasonMessage:
          type: string
          maxLength: 1024
        latitude:
          type: string
          maxLength: 32
        longitude:
          type: string
          maxLength: 32
...
    ShipmentResponse:
      type: object
      required:
        - trackingNo
        - recipient
        - address
        - phone
        - mobile
        - zipCode
        - city
        - district
        - status
      properties:
        trackingNo:
          type: string
          maxLength: 32
        recipient:
          type: string
          maxLength: 128
        address:
          type: string
          maxLength: 512
        phone:
          type: string
          maxLength: 32
        mobile:
          type: string
          maxLength: 32
        zipCode:
          type: string
          maxLength: 16
        city:
          type: string
          maxLength: 64
        district:
          type: string
          maxLength: 64
        status:
          type: string
          maxLength: 64
        signedAt:
          type: string
          nullable: true
          maxLength: 40
        signedImageFileName:
          type: string
          nullable: true
          maxLength: 255
        signedLocation:
          type: string
          nullable: true
          maxLength: 64
    CreateReservationRequest:
      type: object
      required: [address, shipmentNos]
      properties:
        areaCode:
          type: string
          maxLength: 64
        address:
          type: string
          maxLength: 512
        shipmentNos:
          type: array
          maxItems: 200
          items:
            type: string
            maxLength: 64
        fee:
          type: number
        note:
          type: string
          maxLength: 1024
    ReservationResponse:
      type: object
      required: [reservationNo, address, shipmentNos, mode]
      properties:
        reservationNo:
          type: string
          maxLength: 64
        address:
          type: string
          maxLength: 512
        fee:
          type: number
          nullable: true
        shipmentNos:
          type: array
          maxItems: 200
          items:
            type: string
            maxLength: 64
        mode:
          type: string
          maxLength: 16
          enum: [standard, bulk]
    ReservationSupportListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReservationSupportItem'
    ReservationSupportItem:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
          maxLength: 64
        name:
          type: string
          maxLength: 128
        status:
          type: string
          nullable: true
          maxLength: 64
        service:
          type: string
          nullable: true
          maxLength: 128
        role:
          type: string
          nullable: true
          maxLength: 64
        message:
          type: string
          nullable: true
          maxLength: 1024
        reservationNo:
          type: string
          nullable: true
          maxLength: 64
        trackingNo:
          type: string
          nullable: true
          maxLength: 64
        zip:
          type: string
          nullable: true
          maxLength: 64
        areaCode:
          type: string
          nullable: true
          maxLength: 64
        address:
          type: string
          nullable: true
          maxLength: 512
        date:
          type: string
          nullable: true
          maxLength: 40
    RegisterPushTokenRequest:
      type: object
      required: [deviceId, platform, fcmToken]
      properties:
        deviceId:
          type: string
          maxLength: 64
        platform:
          type: string
          maxLength: 16
          enum: [android, ios]
        fcmToken:
          type: string
          maxLength: 4096
        appVersion:
          type: integer
          minimum: 0
    UnregisterPushTokenRequest:
      type: object
      required: [fcmToken]
      properties:
        fcmToken:
          type: string
          maxLength: 4096
...